+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version latest (0.15.1rc?, 2019-09-24?)
 - bugfix: redshift cost optimize total number analyzed was always 0. Fixed to use context variable
 - bugfix: `isitfit cost optimize --n=1` was not breaking early properly. fixed
 - bugfix: ec2 reporters: cleaned up usage of `n_{ec2,rc}_{total,analysed}` and `region_include`; calculating `n_*_analysed` in the analyzer; dropping `n_ec2_analysed` from mainManager
+- bugfix: when datadog is not configured, the `per_ec2` listener should return the context as is, not None, thus breaking the chain for the `optimizerListener`
+- enh: reporters: clean up usage of `analyzer` and `mainManager` in reporters
 
 
 Version 0.15.0 (2019-11-19)